Little code cleanups

+++ database-service/src/main/java/com/example/database_service/repository/ServiceRepository.java
@@ -3,5 +3,5 @@ package com.example.database_service.repository;
 import com.example.database_service.entity.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Service, Integer> {
+public interface ServiceRepository extends JpaRepository<Service, Long> {
 }
